+++ PedidoCompra/PedidoCompra/Domain/Pedidos/Commands/PedidoItem/PedidoItemCommand.cs
@@ -6,7 +6,7 @@ namespace PedidoCompra.Domain.Pedidos.Commands.PedidoItem
     {
         public Guid Id { get; protected set; }
         public string Descricao { get; protected set; }
-        public float Quantidade { get; protected set; }
+        public decimal Quantidade { get; protected set; }
         public decimal ValorUnitario { get; protected set; }
         public Guid PedidoId { get; protected set; }
     }
